Added a SideBar.
The side bar is where the user will go to select blocks to build with.
It is the primary menu used during gameplay. It will slide in when the
user hits tab or presses an arrow on the side of the screen. So far it
animates the slide but I should probably swap to a border layout for the
localizing and resizing of the component.